refactor(dashboard): extract spotlight card component and split recommended cafes
- Create reusable SpotlightCard component for both spots and cafes
- Move RecommendedCafes to separate component file
- Update SuggestedSpots to use new SpotlightCard component
