+++ src/app/dashboard/page.tsx
@@ -3,7 +3,8 @@
 import Image from "next/image"
 import sampleprofile from "../../../public/images/sampleprofile.png"
 import Sidebar from "../../components/Sidebar"
-import SuggestedSpots, { RecommendedCafes } from "./components/SuggestedSpots"
+import SuggestedSpots from "./components/SuggestedSpots"
+import RecommendedCafes from "./components/RecommendedCafes"
 import { useRouter, useSearchParams } from "next/navigation"
 import { useEffect, useState, Suspense } from "react"
 import { useSession } from "next-auth/react"
